add coming soon, update version, show label bottom nav
